+++ packages/clients/src/api/block/v1/types.gen.ts
@@ -284,23 +284,6 @@ export type ImportSnapshotFromObjectStorageRequest = {
   size?: number
 }
 
-export type ImportSnapshotFromS3Request = {
-  /** Zone to target. If none is passed will use default zone from the config. */
-  zone?: ScwZone
-  /** Scaleway Object Storage bucket where the object is stored. */
-  bucket: string
-  /** The object key inside the given bucket. */
-  key: string
-  /** Name of the snapshot. */
-  name: string
-  /** UUID of the Project to which the volume and the snapshot belong. */
-  projectId?: string
-  /** List of tags assigned to the snapshot. */
-  tags?: string[]
-  /** Size of the snapshot. */
-  size?: number
-}
-
 export type ListSnapshotsRequest = {
   /** Zone to target. If none is passed will use default zone from the config. */
   zone?: ScwZone