+++ packages/clients/src/api/block/v1/validation-rules.gen.ts
@@ -1,7 +1,25 @@
 // This file was automatically generated. DO NOT EDIT.
 // If you have any remark or suggestion do not hesitate to open an issue.
 
-export const ClassicMailApiListMailboxesRequest = {
+export const CreateSnapshotRequest = {
+  name: {
+    minLength: 1,
+  },
+}
+
+export const CreateVolumeRequest = {
+  name: {
+    minLength: 1,
+  },
+}
+
+export const ImportSnapshotFromObjectStorageRequest = {
+  name: {
+    minLength: 1,
+  },
+}
+
+export const ListSnapshotsRequest = {
   page: {
     greaterThan: 0,
   },
@@ -11,7 +29,7 @@ export const ClassicMailApiListMailboxesRequest = {
   },
 }
 
-export const ListControlPanelsRequest = {
+export const ListVolumeTypesRequest = {
   page: {
     greaterThan: 0,
   },
@@ -21,12 +39,18 @@ export const ListControlPanelsRequest = {
   },
 }
 
-export const ListHostingsRequest = {
+export const ListVolumesRequest = {
   page: {
     greaterThan: 0,
   },
   pageSize: {
     greaterThan: 0,
     lessThanOrEqual: 100,
+  },
+}
+
+export const Reference = {
+  productResourceType: {
+    minLength: 1,
   },
 }