+++ packages/clients/src/api/cockpit/v1/marshalling.gen.ts
@@ -22,11 +22,11 @@ import type {
   Grafana,
   GrafanaProductDashboard,
   GrafanaUser,
+  ListAlertsResponse,
   ListContactPointsResponse,
   ListDataSourcesResponse,
   ListGrafanaProductDashboardsResponse,
   ListGrafanaUsersResponse,
-  ListManagedAlertsResponse,
   ListPlansResponse,
   ListTokensResponse,
   Plan,
@@ -236,6 +236,40 @@ export const unmarshalGrafana = (data: unknown): Grafana => {
   } as Grafana
 }
 
+const unmarshalAlert = (data: unknown): Alert => {
+  if (!isJSONObject(data)) {
+    throw new TypeError(
+      `Unmarshalling the type 'Alert' failed as data isn't a dictionary.`,
+    )
+  }
+
+  return {
+    annotations: data.annotations,
+    duration: data.duration,
+    enabled: data.enabled,
+    name: data.name,
+    preconfigured: data.preconfigured,
+    region: data.region,
+    rule: data.rule,
+    state: data.state ? data.state : undefined,
+  } as Alert
+}
+
+export const unmarshalListAlertsResponse = (
+  data: unknown,
+): ListAlertsResponse => {
+  if (!isJSONObject(data)) {
+    throw new TypeError(
+      `Unmarshalling the type 'ListAlertsResponse' failed as data isn't a dictionary.`,
+    )
+  }
+
+  return {
+    alerts: unmarshalArrayOfObject(data.alerts, unmarshalAlert),
+    totalCount: data.total_count,
+  } as ListAlertsResponse
+}
+
 export const unmarshalListContactPointsResponse = (
   data: unknown,
 ): ListContactPointsResponse => {
@@ -305,37 +339,6 @@ export const unmarshalListGrafanaUsersResponse = (
     ),
     totalCount: data.total_count,
   } as ListGrafanaUsersResponse
-}
-
-const unmarshalAlert = (data: unknown): Alert => {
-  if (!isJSONObject(data)) {
-    throw new TypeError(
-      `Unmarshalling the type 'Alert' failed as data isn't a dictionary.`,
-    )
-  }
-
-  return {
-    description: data.description,
-    name: data.name,
-    product: data.product,
-    productFamily: data.product_family,
-    rule: data.rule,
-  } as Alert
-}
-
-export const unmarshalListManagedAlertsResponse = (
-  data: unknown,
-): ListManagedAlertsResponse => {
-  if (!isJSONObject(data)) {
-    throw new TypeError(
-      `Unmarshalling the type 'ListManagedAlertsResponse' failed as data isn't a dictionary.`,
-    )
-  }
-
-  return {
-    alerts: unmarshalArrayOfObject(data.alerts, unmarshalAlert),
-    totalCount: data.total_count,
-  } as ListManagedAlertsResponse
 }
 
 export const unmarshalListPlansResponse = (