+++ packages/clients/src/api/webhosting/index.gen.ts
@@ -4,4 +4,3 @@
  */
 
 export * as v1 from './v1/index.gen'
-export * as v1alpha1 from './v1alpha1/index.gen'