+++ packages/clients/src/api/webhosting/v1/marshalling.gen.ts
@@ -167,6 +167,7 @@ export const unmarshalDnsRecords = (data: unknown): DnsRecords => {
   }
 
   return {
+    dnsConfig: data.dns_config,
     nameServers: unmarshalArrayOfObject(data.name_servers, unmarshalNameserver),
     records: unmarshalArrayOfObject(data.records, unmarshalDnsRecord),
     status: data.status,
@@ -298,8 +299,9 @@ export const unmarshalHosting = (data: unknown): Hosting => {
 
   return {
     createdAt: unmarshalDate(data.created_at),
-    dnsStatus: data.dns_status,
+    dnsStatus: data.dns_status ? data.dns_status : undefined,
     domain: data.domain,
+    domainStatus: data.domain_status,
     id: data.id,
     ipv4: data.ipv4,
     offer: data.offer ? unmarshalOffer(data.offer) : undefined,
@@ -401,8 +403,9 @@ const unmarshalHostingSummary = (data: unknown): HostingSummary => {
 
   return {
     createdAt: unmarshalDate(data.created_at),
-    dnsStatus: data.dns_status,
+    dnsStatus: data.dns_status ? data.dns_status : undefined,
     domain: data.domain,
+    domainStatus: data.domain_status,
     id: data.id,
     offerName: data.offer_name,
     projectId: data.project_id,