+++ packages/clients/src/api/webhosting/v1/types.gen.ts
@@ -42,6 +42,7 @@ export type DomainDnsAction =
   | 'auto_config_web_records'
   | 'auto_config_mail_records'
   | 'auto_config_nameservers'
+  | 'auto_config_none'
 
 export type DomainStatus =
   | 'unknown_status'
@@ -58,15 +59,6 @@ export type DomainZoneOwner =
   | 'webhosting'
 
 export type HostingStatus =
-  | 'unknown_status'
-  | 'delivering'
-  | 'ready'
-  | 'deleting'
-  | 'error'
-  | 'locked'
-  | 'migrating'
-
-export type HostingSummaryStatus =
   | 'unknown_status'
   | 'delivering'
   | 'ready'
@@ -302,15 +294,17 @@ export interface HostingSummary {
   /** Date on which the Web Hosting plan was last updated. */
   updatedAt?: Date
   /** Status of the Web Hosting plan. */
-  status: HostingSummaryStatus
+  status: HostingStatus
   /** Main domain associated with the Web Hosting plan. */
   domain: string
   /** Whether the hosting is protected or not. */
   protected: boolean
-  /** DNS status of the Web Hosting plan. */
-  dnsStatus: DnsRecordsStatus
+  /** @deprecated DNS status of the Web Hosting plan. */
+  dnsStatus?: DnsRecordsStatus
   /** Name of the active offer for the Web Hosting plan. */
   offerName: string
+  /** Main domain status of the Web Hosting plan. */
+  domainStatus: DomainStatus
   /** Region where the Web Hosting plan is hosted. */
   region: ScwRegion
 }
@@ -611,6 +605,8 @@ export interface DnsRecords {
   nameServers: Nameserver[]
   /** Status of the records. */
   status: DnsRecordsStatus
+  /** Records dns auto configuration settings. */
+  dnsConfig: DomainDnsAction[]
 }
 
 export interface Domain {
@@ -708,14 +704,16 @@ export interface Hosting {
   platform?: Platform
   /** List of tags associated with the Web Hosting plan. */
   tags: string[]
-  /** DNS status of the Web Hosting plan. */
-  dnsStatus: DnsRecordsStatus
+  /** @deprecated DNS status of the Web Hosting plan. */
+  dnsStatus?: DnsRecordsStatus
   /** Current IPv4 address of the hosting. */
   ipv4: string
   /** Whether the hosting is protected or not. */
   protected: boolean
   /** Details of the hosting user. */
   user?: HostingUser
+  /** Main domain status of the Web Hosting plan. */
+  domainStatus: DomainStatus
   /** Region where the Web Hosting plan is hosted. */
   region: ScwRegion
 }