+++ packages/clients/src/scw/constants.ts
@@ -1,3 +1,3 @@
-export const version = 'v2.65.0'
+export const version = 'v2.68.0'
 
 export const userAgent = `scaleway-sdk-js/${version}`