+++ packages/clients/CHANGELOG.md
@@ -3,6 +3,53 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.68.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk@2.67.0...@scaleway/sdk@2.68.0) (2025-03-03)
+
+### Features
+
+- **audit_trail:** add organization_id field to ListProducts ([#1853](https://github.com/scaleway/scaleway-sdk-js/issues/1853)) ([1c5ada4](https://github.com/scaleway/scaleway-sdk-js/commit/1c5ada42e2fb6c391545063d2daa64211a83713c))
+- **audit_trail:** add support for `Event.resources` ([#1849](https://github.com/scaleway/scaleway-sdk-js/issues/1849)) ([63c10bd](https://github.com/scaleway/scaleway-sdk-js/commit/63c10bd50419f4ade54d7ed515e4f255ec257d6d))
+- **audit_trail:** migrate iamx to iam ([#1834](https://github.com/scaleway/scaleway-sdk-js/issues/1834)) ([e511673](https://github.com/scaleway/scaleway-sdk-js/commit/e511673335fe2d3f9d558a00019908f324472934))
+- **baremetal:** add support for `MigrateServerToMonthlyOfferRequest` ([#1836](https://github.com/scaleway/scaleway-sdk-js/issues/1836)) ([9f18271](https://github.com/scaleway/scaleway-sdk-js/commit/9f182712c3dce2e8c87fead349db2fed6483fe24))
+- **block:** add support for v1 ([#1855](https://github.com/scaleway/scaleway-sdk-js/issues/1855)) ([3b16b6f](https://github.com/scaleway/scaleway-sdk-js/commit/3b16b6fe00efed25fb73e7008f3f931871a9bc8f))
+- **block:** add tags filter in list snapshots ([#1832](https://github.com/scaleway/scaleway-sdk-js/issues/1832)) ([b7654e6](https://github.com/scaleway/scaleway-sdk-js/commit/b7654e65f1b59efab8bf1237393c6d6b757fe9da))
+- **edge_services:** add support for v1beta1 ([#1833](https://github.com/scaleway/scaleway-sdk-js/issues/1833)) ([0040898](https://github.com/scaleway/scaleway-sdk-js/commit/004089827181764ddb2db734bdd58df0c849d2c9))
+- **instance:** add compatible types endpoint and end_of_service flag in ServerType ([#1835](https://github.com/scaleway/scaleway-sdk-js/issues/1835)) ([f849b3e](https://github.com/scaleway/scaleway-sdk-js/commit/f849b3e8b195b70ee14a0b6ab528ce131c1c543b))
+- **lb:** add support for `matchSubdomains` ([#1864](https://github.com/scaleway/scaleway-sdk-js/issues/1864)) ([c8e36ae](https://github.com/scaleway/scaleway-sdk-js/commit/c8e36ae0c698cee55a64050148ef6d1248c2031b))
+- **mongodb:** implement roles permissions api ([#1857](https://github.com/scaleway/scaleway-sdk-js/issues/1857)) ([4c86fbd](https://github.com/scaleway/scaleway-sdk-js/commit/4c86fbd890c8d8612507da1e8a4e1ce6922b525d))
+
+### Bug Fixes
+
+- **webhosting:** add hosting status naming ([#1859](https://github.com/scaleway/scaleway-sdk-js/issues/1859)) ([be04eda](https://github.com/scaleway/scaleway-sdk-js/commit/be04edadb76ec6116138b85c691e91d97f0c2275))
+
+## [2.67.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk@2.66.0...@scaleway/sdk@2.67.0) (2025-02-24)
+
+### Features
+
+- **audit_trail:** add IAM resources ([#1819](https://github.com/scaleway/scaleway-sdk-js/issues/1819)) ([c5e65f1](https://github.com/scaleway/scaleway-sdk-js/commit/c5e65f14b032c8ee62ad15b504345d381a4e2bc3))
+- **cockpit:** add support for `RegionalApiListAlertsRequest` ([#1806](https://github.com/scaleway/scaleway-sdk-js/issues/1806)) ([bb46893](https://github.com/scaleway/scaleway-sdk-js/commit/bb468930776ca2459f040040d05c292cffc2b5ef))
+- **lb:** add support for `connectionRateLimit` ([#1811](https://github.com/scaleway/scaleway-sdk-js/issues/1811)) ([4997495](https://github.com/scaleway/scaleway-sdk-js/commit/4997495752cc405cf91f5915b7b6232cc0f493aa))
+- **secret:** add support for `RestoreSecretRequest` and `RestoreSecretVersionRequest` ([#1826](https://github.com/scaleway/scaleway-sdk-js/issues/1826)) ([eaf54ab](https://github.com/scaleway/scaleway-sdk-js/commit/eaf54ab8b6463c8a464ec4ce82ccd46eb2813a74))
+- **vpc_gw:** add a call to migrate a V1 gateway to V2 ([#1827](https://github.com/scaleway/scaleway-sdk-js/issues/1827)) ([a8f5160](https://github.com/scaleway/scaleway-sdk-js/commit/a8f51604b5ac5572a20f7e71782f04dc3ee19a86))
+- **webhosting:** add domain selected dns configuration ([#1831](https://github.com/scaleway/scaleway-sdk-js/issues/1831)) ([5327dc1](https://github.com/scaleway/scaleway-sdk-js/commit/5327dc1e19a8055ac11cbb481de0e625854fb219))
+
+### Bug Fixes
+
+- **webhosting:** add hosting domain status ([#1829](https://github.com/scaleway/scaleway-sdk-js/issues/1829)) ([44184c2](https://github.com/scaleway/scaleway-sdk-js/commit/44184c28886d2dbdddec426ba5b4cf7951c6e5b2))
+
+## [2.66.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk@2.65.0...@scaleway/sdk@2.66.0) (2025-02-17)
+
+### Features
+
+- **apple_silicon:** introduce monthly commitment handling ([#1803](https://github.com/scaleway/scaleway-sdk-js/issues/1803)) ([0972b37](https://github.com/scaleway/scaleway-sdk-js/commit/0972b37b4592b0112e6f800426b8129a85192d76))
+- **audit_trail:** add key manager to resource api ([#1786](https://github.com/scaleway/scaleway-sdk-js/issues/1786)) ([b985b98](https://github.com/scaleway/scaleway-sdk-js/commit/b985b981e60650c72ed4b8faf2422e3cf4063f08))
+- **edge_services:** add WAF billing ([#1800](https://github.com/scaleway/scaleway-sdk-js/issues/1800)) ([84a9ec5](https://github.com/scaleway/scaleway-sdk-js/commit/84a9ec517b24259cf6beef3277790a4764750bbb))
+
+### Bug Fixes
+
+- add prefix to Zone and Region import ([#1802](https://github.com/scaleway/scaleway-sdk-js/issues/1802)) ([ce59650](https://github.com/scaleway/scaleway-sdk-js/commit/ce596501b569b3387a76e12e7d00a90d94ca803e))
+- **types:** unknown property accessKey in createClient ([#1781](https://github.com/scaleway/scaleway-sdk-js/issues/1781)) ([865b8c7](https://github.com/scaleway/scaleway-sdk-js/commit/865b8c7b6854314ebc209b5bba6ea836cb792b73))
+
 ## [2.65.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk@2.64.0...@scaleway/sdk@2.65.0) (2025-02-10)
 
 ### Features